requirements.txt
snowflake-snowpark-python[pandas]
tomli
toml
pytest
snowflake-vcrpy @ git+https://github.com/Snowflake-Labs/snowflake-vcrpy.git@v0.1.1
streamlit
plotly